requirements.txt
# ============================================
# Multi-Agent 深度搜索系统依赖包
# ============================================

# ============================================
# 核心框架
# ============================================
# LangChain 核心组件
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langgraph-checkpoint-postgres>=1.0.0

# LlamaIndex 核心库和扩展
llama-index>=0.9.0
llama-index-core>=0.9.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-readers-file>=0.1.0  # MarkdownReader

# ============================================
# LLM 和 Embedding
# ============================================
# OpenAI 兼容接口（支持 vllm 等本地部署）
openai>=1.0.0

# ============================================
# 向量数据库
# ============================================
# Chroma 向量数据库
chromadb>=0.4.0

# ============================================
# 数据库
# ============================================
# PostgreSQL 异步驱动（用于短期记忆）
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# ============================================
# Reranker
# ============================================
# TEI 客户端（用于调用 BGE Reranker）
tei-client>=0.1.0

# FlagEmbedding（可选，用于本地部署 BGE Reranker）
FlagEmbedding>=1.2.0

# ============================================
# Web 框架
# ============================================
# FastAPI 和 Uvicorn
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ============================================
# HTTP 客户端
# ============================================
# 异步 HTTP 客户端
httpx>=0.24.0

# ============================================
# 工具和搜索
# ============================================
# Tavily 搜索
tavily-python>=0.3.0

# MCP 适配器
langchain-mcp-adapters>=0.1.0

# 异步文件操作
aiofiles>=23.0.0

# Wikipedia 搜索
wikipedia>=1.4.0

# XML 解析（用于 PubMed）
xmltodict>=0.13.0

# AkShare（金融数据）
akshare>=1.12.0

# ============================================
# 文档处理
# ============================================
# MinerU 客户端（用于 PDF 转 Markdown）
mineru-vl-utils>=0.1.0

# Markdown 处理
markdown>=3.5.0

# ============================================
# 日志
# ============================================
# 并发日志处理
concurrent-log-handler>=0.9.20

# ============================================
# 环境变量
# ============================================
# 环境变量管理
python-dotenv>=1.0.0

# ============================================
# 测试框架（可选，见 requirements-test.txt）
# ============================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# hypothesis>=6.82.0

# ============================================
# 其他工具
# ============================================
# 类型检查
typing-extensions>=4.8.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# JSON 处理
orjson>=3.9.0

# 日期时间处理
python-dateutil>=2.8.0

# ============================================
# 可选依赖
# ============================================
# vllm（用于本地部署 Embedding 模型，需要 GPU）
# vllm>=0.2.0

# Gunicorn（生产环境部署）
# gunicorn>=21.2.0

# Prometheus 监控（可选）
# prometheus-client>=0.19.0

# ============================================
# 版本说明
# ============================================
# Python 版本要求: >=3.10
# 安装命令（使用清华镜像源）:
#   pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt
#
# 注意事项:
# 1. vllm 需要 GPU 支持，如果没有 GPU 可以不安装
# 2. FlagEmbedding 用于本地部署 Reranker，如果使用 TEI 服务可以不安装
# 3. magic-pdf (MinerU) 可能需要额外的系统依赖
# 4. 测试依赖请参考 requirements-test.txt
